Refactoring and code clean up nased on review comments
