+++ src/main/java/uk/gov/ch/pocappointmentapi/AppointmentController.java
@@ -7,20 +7,22 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
+import uk.gov.ch.pocappointmentapi.domain.Appointment;
+
 @RestController
 public class AppointmentController {
 
-	@Autowired
-	AppointmentRepository repository;
+    @Autowired
+    AppointmentRepository repository;
 
-	@GetMapping(value = "/company/{companyNumber}/appointment/{appointmentId}")
-	public Appointment appointment(@PathVariable String companyNumber, @PathVariable String appointmentId) {
-		return repository.findByCompanyNumberAndAppointmentId(companyNumber, appointmentId);
-	}
+    @GetMapping(value = "/company/{companyNumber}/appointment/{appointmentId}")
+    public Appointment appointment(@PathVariable String companyNumber, @PathVariable String appointmentId) {
+        return repository.findByCompanyNumberAndAppointmentId(companyNumber, appointmentId);
+    }
 
-	@GetMapping(value = "/healthcheck")
-	@ResponseStatus(HttpStatus.OK)
-	public String healthcheck() {
-		return "Appointment API is available";
-	}
+    @GetMapping(value = "/healthcheck")
+    @ResponseStatus(HttpStatus.OK)
+    public String healthcheck() {
+        return "Appointment API is available";
+    }
 }
