+++ src/main/java/uk/gov/ch/pocappointmentapi/AppointmentRepository.java
@@ -2,8 +2,10 @@ package uk.gov.ch.pocappointmentapi;
 
 import org.springframework.data.mongodb.repository.MongoRepository;
 
+import uk.gov.ch.pocappointmentapi.domain.Appointment;
+
 public interface AppointmentRepository extends MongoRepository<Appointment, String> {
-	
-	Appointment findByCompanyNumberAndAppointmentId(String companyNumber, String appointmentId);
+
+    Appointment findByCompanyNumberAndAppointmentId(String companyNumber, String appointmentId);
 
 }
