+++ src/test/java/uk/gov/ch/pocappointmentapi/PocAppointmentApiTests.java
@@ -9,8 +9,8 @@ import org.springframework.test.context.junit4.SpringRunner;
 @SpringBootTest
 public class PocAppointmentApiTests {
 
-	@Test
-	public void contextLoads() {
-	}
+    @Test
+    public void contextLoads() {
+    }
 
 }
